+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 * Add `generate_systemd_units` method to `Pod` and `Container`
 * Add `generate_kube_yaml` method to `Pod` and `Container`
 * Add `Podman::play_kubernetes_yaml`
+* Add `Podman::remove_kubernetes_pods`